Completed assignment

+++ app/controllers/toys_controller.rb
@@ -7,19 +7,32 @@ class ToysController < ApplicationController
   end
 
   def create
-    toy = Toys.create(toy_params)
-    render json: toy, status: :created
+    toy = Toy.create(toy_params)
+    if toy.valid?
+      render json: toy, status: :created
+    else
+      render json: {error: toy.errors.full_messages}, status: :unprocessable_entity
+    end
   end
 
   def update
     toy = Toy.find_by(id: params[:id])
-    toy.update(toy_params)
+    if toy
+      toy.update(likes: toy.likes + 1)
+      render json: toy, status: :ok
+    else
+      render json: {error: "Toy not found"}, status: :not_found
+    end
   end
 
   def destroy
     toy = Toy.find_by(id: params[:id])
-    toy.destroy
-    head :no_content
+    if toy
+      toy.destroy
+      head :no_content
+    else
+      render json: {error: "Toy not found"}, status: :not_found
+    end
   end
 
   private
